cobra-ify the operator and configurator
This commit wraps the operator (manager) and configurator into a single
redpanda-operator binary that utilizes cobra to parse CLI flags. As there are
many ways to deploy the operator and configurator, backwards compatibility is
provided by a small `alias` program that occupies the original `/manager` and
`/configurator` binaries has been added to the containers.

This is done in preparation for migrating the helm chart's jobs into go code
that will be shared by the operator and the helm chart while flux is being
removed.

Note: the original `cmd/main.go` was moved with very minor modifications to
`cmd/run/run.go`. Due to `cmd/main.go` becoming a new file and the naming
conventions of go/cobra, git is not detecting the move correctly.
