+++ alias/go.mod
@@ -1,0 +1,3 @@
+module github.com/redpanda-data/redpanda-operator/alias
+
+go 1.22.4
