+++ src/go/k8s/cmd/configurator/configurator_test.go
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-package main
+package configurator
 
 import (
 	"testing"
