+++ taskfiles/goreleaser.yml
@@ -45,4 +45,4 @@ tasks:
       - task: :dev:install-goreleaser
       - task: :k8s:generate
     cmds:
-      - '{{.GORELEASER_INSTALL_DIR}}/goreleaser build --snapshot --clean --id configurator --id redpanda-operator'
+      - '{{.GORELEASER_INSTALL_DIR}}/goreleaser build --snapshot --clean'
